+++ packages/client/src/hooks/queries/comment/useDeleteComment.ts
@@ -1,4 +1,4 @@
-import CommentAPI from '~/lib/api/comment';
+import { CommentAPI } from '~/lib/api';
 import { useMutation } from '@tanstack/react-query';
 import type { UseMutationOptionsOf } from '~/hooks/queries/types';
