+++ packages/client/src/hooks/queries/post/useGetSearchPosts.ts
@@ -1,4 +1,4 @@
-import PostAPI from '~/lib/api/post';
+import { PostAPI } from '~/lib/api';
 import { useQuery } from '@tanstack/react-query';
 import type { UseQueryOptionsOf } from '~/hooks/queries/types';
