+++ packages/client/src/hooks/queries/series/useAppendPostToSeries.ts
@@ -1,4 +1,4 @@
-import SeriesAPI from '~/lib/api/series';
+import { SeriesAPI } from '~/lib/api';
 import { useMutation } from '@tanstack/react-query';
 import type { UseMutationOptionsOf } from '~/hooks/queries/types';
