+++ packages/client/src/hooks/queries/user/useGetUserByUsername.ts
@@ -1,4 +1,4 @@
-import UserAPI from '~/lib/api/user';
+import { UserAPI } from '~/lib/api';
 import { useQuery } from '@tanstack/react-query';
 import type { UseQueryOptionsOf } from '~/hooks/queries/types';
