+++ packages/client/src/lib/api/apiClient.ts
@@ -1,6 +1,8 @@
 import axios from 'axios';
+import { BASE_URL } from '~/constants';
 
 const apiClient = axios.create({
+  baseURL: BASE_URL,
   withCredentials: true,
 });
