+++ packages/client/src/lib/api/auth.api.ts
@@ -1,11 +1,9 @@
 import { API } from '~/constants';
 import apiClient from './apiClient';
 
-const AuthAPI = {
+export const AuthAPI = {
   logout: async () => {
     const { data } = await apiClient.delete(`${API.AUTH}/logout`);
     return data;
   },
 };
-
-export default AuthAPI;
